filename cleanup
