--- handbook-shacl.ttl
@@ -1,72 +1,0 @@
-@prefix code: <https://uwa.handbook/code/> .
-@prefix rel: <https://uwa.handbook/relation/> .
-@prefix uwa: <https://uwa.handbook/> .
-@prefix schema: <http://schema.org/> .
-@prefix sh: <http://www.w3.org/ns/shacl#> .
-@prefix xsd: <http://www.w3.org/2001/XMLSchema#> .
-
-schema:PrereqShape a sh:NodeShape ;
-    sh:targetClass uwa:unit ;
-    sh:message "Prerequisite must have a lower unit level than unit" ;
-    sh:or (
-        [
-            sh:property [
-                sh:path rel:Level ;
-                sh:hasValue "1"^^xsd:integer ;
-            ]
-        ]
-        [
-            sh:sparql [
-                    a sh:SPARQLConstraint ;
-                    sh:message "Prerequisite must have a lower unit level than unit" ;
-                    sh:prefix rel: ;
-                    sh:prefix uwa: ;
-                    sh:select """
-                        SELECT $this
-                        WHERE {
-                            $this rel:Level ?level .
-                            $this rel:Prerequisite ?prereq .
-                            ?prereq a uwa:unit .
-                            ?prereq rel:Level ?prereqLevel
-                            FILTER (?prereqLevel > ?level)
-                        }
-                    """ ;
-                ] 
-        ]
-    ) .
-    
-schema:SelfPrereqShape a sh:NodeShape ;
-    sh:targetClass uwa:unit ;
-    sh:sparql [
-        a sh:SPARQLConstraint ;
-        sh:message "Unit cannot be a prerequisite of itself." ;
-        sh:prefixes rel: ;
-        sh:select """
-            SELECT $this
-            WHERE {
-                $this rel:Prerequisite $this .
-            }
-        """ ;
-    ] .
-
-schema:ContactShape a sh:NodeShape ;
-    sh:targetClass uwa:major ;
-    sh:sparql [
-        a sh:SPARQLConstraint ;
-        sh:message "Major has more than 40 Contact Hours per Week" ;
-        sh:prefix rel: ;
-        sh:prefix uwa: ;
-        sh:select """
-            SELECT $this
-            WHERE {
-                $this a uwa:major.
-                OPTIONAL { $this rel:L4Unit ?any }
-                BIND(IF(bound(?any), 8, 6) AS ?semesters)
-                $this ?level ?unit.
-                ?unit a uwa:unit.
-                ?unit rel:ContactHours ?hours.
-            }
-            GROUP BY $this
-            HAVING (SUM (?hours) / ?semesters > 40)
-        """ ;
-    ] .
